LUS-11785 libcfs: introduce genradix support
For a long time it has been known that vmalloc allocation have
a performance penalty. So another solution was introduced using a
generic radix tree to manage page size allocations. This new
functionality has huge potential to give us large performance gains.
Also this new API has an advantage over vmalloc for the case of not
knowing ahead of time how much to allocate. This is the case of a
Netlink data stream that is sending a array of data. We can use this
for the case of needing to grow an array instead of guessing how big
the data is coming.

Lustre-change: https://review.whamcloud.com/45890
Lustre-commit: 841843f44cd16e2bfb9b31195f7398cff02f9088

Test-Parameters: trivial
Change-Id: I2d7b78569defd651fbb75bfbd82e0d805aff436e
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162499
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>