LUS-11633 mgc: reconnection without pinger
When MGS was offline for some time, AT is increased and
connection request deadline is high. Reconnect with a pinger
waits a request deadline for a next attempt. A situation is
worse with a failover partner, when different connections are used.
Reconnection could fail with local MGS too. The patch force
reconnection with import invalidate and abort inflight requests.

ptlrpc_recover_import() aborts waiting for disconnect import state.
But disconnect happens between connection attempt and it is valid.
This is fixed.

Reset Adaptive Timeout when local MGS starts. It allows MGC to
reconnect efficiently.

Signed-off-by: Alexander Boyko <alexander.boyko@hpe.com>
Change-Id: I5987160a95337b1e608ce8999cb3be5db9ee344d
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162403
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>