LUS-11886 build: The alloc_inode_sb check needs to be run
Unfortunately the patch that includes the alloc_inode_sb
does not have the parallel build infrastructure that in
needs to run.

There is not catchall LB2_LINUX_TEST_COMPILE_ALL so
include it in the LC_PROG_LINUX_SRC macro.

This is only needed until the parallel build infrastructure
is landed to cray-2.15

Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I513415fd71a2d3936633a62663ad51fedd77e89f